requirements.txt
asgiref==3.7.2
Django==4.2.4
django-cors==0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
pillow==10.2.0
PyJWT==2.8.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
